+++ install/install-peco.sh
@@ -1,3 +1,8 @@
+if [ `uname` = "Darwin" ]; then
+  echo "install via brew"
+  exit(0)
+fi
+
 rm -rf $DOTFILES/local/peco
 ghbin peco/peco # move into pkg
 mv $DOTFILES/pkg/peco $DOTFILES/local/peco
